Issue in ReportsAuditController

+++ Infin8.Coapp.Dto/rptTermDepositPayable.cs
@@ -5,7 +5,7 @@
         public int TDScheme_Id { get; set; }
         public string? TDScheme_Name { get; set; }
         public string? TDSchemeType { get; set; }
-        public int TD_Id { get; set; }
+        public decimal TD_Id { get; set; }
         public string? TD_No { get; set; }
         public DateTime Trn_Date { get; set; }
         public string? TDH_Name { get; set; }
